Moved Output class to a seperate file
